Implement separate layout saving for mobile and desktop
Updated the layout handling mechanism to differentiate between mobile and desktop layouts. Previously, we used the same layout for both desktop and mobile users - `homeLayout`. Now, we have introduced two more layouts - `mobileHomeLayout` and `desktopHomeLayout` - that will be used accordingly based on whether the user is on a mobile or desktop. This implementation makes layout more adaptive and user-friendly, tailoring the user interface to the specific device in use.